+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Fixed bug where optimize button doesn't optimize the itinerary
+- Fixed coordinates text input too crowded due to accuracy
 
 ### Changed
 - Changed edit duration modal to use a more ui coherent timer picker