+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed coordinates text input too crowded due to 
 - Fixed bug where admin cannot update places without an error occuring regarding ALL_LANDMARKS
 - Fixed discard and delete dialogs not rounded enough
+- Fixed an issue where admins see an error when creating places regarding "No landmarks found"
 
 ### Added
 - Added fullscreen mode in location selection dialog