+++ CHANGELOG.md
@@ -5,7 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/niceWizzard/thesis-lakad-rn/compare/1.2.1...HEAD)
+## [Unreleased](https://github.com/niceWizzard/thesis-lakad-rn/compare/1.3.0...HEAD)
+
+
+## [1.3.0](https://github.com/niceWizzard/thesis-lakad-rn/compare/1.2.1...1.3.0)
 
 ### Fixed
 - Fixed bug where optimize button doesn't optimize the itinerary